Create index.html
Very first step to creating my personal site. This file is in html and bootstrap 3.

Sets up the structure of the site, including a "jumbotron" with my name, a short bio, and (currently unfinished) social buttons for linkedin and github. The next container is an about me section including what is currently filler text.

This is currently unstyled and content is incomplete.